+++ packages/ethereum-contracts/contracts/agreements/ConstantFlowAgreementV1.sol
@@ -1048,6 +1048,7 @@ contract ConstantFlowAgreementV1 is
     // solhint-disable-next-line contract-name-camelcase
     struct _StackVars_changeFlowToApp {
         bytes cbdata;
+        uint256 callbackGasLimit;
         FlowData newFlowData;
         ISuperfluid.Context appContext;
     }
@@ -1084,7 +1085,7 @@ contract ConstantFlowAgreementV1 is
             } else /* if (optype == FlowChangeType.DELETE_FLOW) */ {
                 cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_TERMINATED_NOOP;
             }
-            vars.cbdata = AgreementLibrary.callAppBeforeCallback(cbStates, ctx);
+            (vars.cbdata, vars.callbackGasLimit) = AgreementLibrary.callAppBeforeCallback(cbStates, ctx);
 
             ISuperfluidGovernance gov = ISuperfluidGovernance(ISuperfluid(msg.sender).getGovernance());
 
@@ -1122,7 +1123,8 @@ contract ConstantFlowAgreementV1 is
             } else /* if (optype == FlowChangeType.DELETE_FLOW) */ {
                 cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_TERMINATED_NOOP;
             }
-            (vars.appContext, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+            (vars.appContext, newCtx) = AgreementLibrary.callAppAfterCallback(
+                cbStates, vars.cbdata, vars.callbackGasLimit, newCtx);
 
             // NB: the callback might update the same flow!!
             // reload the flow data