+++ packages/ethereum-contracts/contracts/agreements/InstantDistributionAgreementV1.sol
@@ -372,6 +372,7 @@ contract InstantDistributionAgreementV1 is
          IndexData idata;
          SubscriptionData sdata;
          bytes cbdata;
+         uint256 callbackGasLimit;
      }
 
     /// @dev IInstantDistributionAgreementV1.approveSubscription implementation
@@ -419,7 +420,7 @@ contract InstantDistributionAgreementV1 is
 
         if (!vars.subscriptionExists) {
             cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_CREATED_NOOP;
-            vars.cbdata = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
+            (vars.cbdata, vars.callbackGasLimit) = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
 
             vars.sdata = SubscriptionData({
                 publisher: publisher,
@@ -433,10 +434,11 @@ contract InstantDistributionAgreementV1 is
             token.createAgreement(vars.sId, _encodeSubscriptionData(vars.sdata));
 
             cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_CREATED_NOOP;
-            (, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+            (, newCtx) = AgreementLibrary.callAppAfterCallback(
+                cbStates, vars.cbdata, vars.callbackGasLimit, newCtx);
         } else {
             cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_UPDATED_NOOP;
-            vars.cbdata = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
+            (vars.cbdata, vars.callbackGasLimit) = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
             // NOTE casting these values to int256 is okay because the original values
             // are uint128
             int balanceDelta = int256(uint256(vars.idata.indexValue - vars.sdata.indexValue))
@@ -456,7 +458,8 @@ contract InstantDistributionAgreementV1 is
             token.updateAgreementData(vars.sId, _encodeSubscriptionData(vars.sdata));
 
             cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_UPDATED_NOOP;
-            (, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+            (, newCtx) = AgreementLibrary.callAppAfterCallback(
+                cbStates, vars.cbdata, vars.callbackGasLimit, newCtx);
         }
 
         // can index up to three words, hence splitting into two events from publisher or subscriber's view.
@@ -505,7 +508,7 @@ contract InstantDistributionAgreementV1 is
         newCtx = ctx;
 
         cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_TERMINATED_NOOP;
-        vars.cbdata = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
+        (vars.cbdata, vars.callbackGasLimit) = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
         // NOTE downcasting these values to int256 is okay because the original values
         // are uint128
         int256 balanceDelta = int256(uint256(vars.idata.indexValue - vars.sdata.indexValue))
@@ -527,7 +530,8 @@ contract InstantDistributionAgreementV1 is
         token.settleBalance(subscriber, balanceDelta);
 
         cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_TERMINATED_NOOP;
-        (, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+        (, newCtx) = AgreementLibrary.callAppAfterCallback(
+            cbStates, vars.cbdata, vars.callbackGasLimit, newCtx);
 
         emit IndexUnsubscribed(token, publisher, indexId, subscriber, userData);
         emit SubscriptionRevoked(token, subscriber, publisher, indexId, userData);
@@ -576,10 +580,10 @@ contract InstantDistributionAgreementV1 is
         // before-hook callback
         if (vars.subscriptionExists) {
             cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_UPDATED_NOOP;
-            vars.cbdata = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
+            (vars.cbdata, vars.callbackGasLimit) = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
         } else {
             cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_CREATED_NOOP;
-            vars.cbdata = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
+            (vars.cbdata, vars.callbackGasLimit) = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
         }
 
         // update publisher data
@@ -643,10 +647,12 @@ contract InstantDistributionAgreementV1 is
         // after-hook callback
         if (vars.subscriptionExists) {
             cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_UPDATED_NOOP;
-            (, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+            (, newCtx) = AgreementLibrary.callAppAfterCallback(
+                cbStates, vars.cbdata, vars.callbackGasLimit, newCtx);
         } else {
             cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_CREATED_NOOP;
-            (, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+            (, newCtx) = AgreementLibrary.callAppAfterCallback(
+                cbStates, vars.cbdata, vars.callbackGasLimit, newCtx);
         }
 
         emit IndexUnitsUpdated(token, publisher, indexId, subscriber, units, userData);
@@ -799,7 +805,7 @@ contract InstantDistributionAgreementV1 is
         newCtx = ctx;
 
         cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_TERMINATED_NOOP;
-        vars.cbdata = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
+        (vars.cbdata, vars.callbackGasLimit) = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
         // NOTE casting these values to int256 is okay because the original values
         // are uint128
         int256 balanceDelta = int256(uint256(vars.idata.indexValue - vars.sdata.indexValue))
@@ -831,7 +837,8 @@ contract InstantDistributionAgreementV1 is
         token.settleBalance(subscriber, balanceDelta);
 
         cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_TERMINATED_NOOP;
-        (, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+        (, newCtx) = AgreementLibrary.callAppAfterCallback(
+            cbStates, vars.cbdata, vars.callbackGasLimit, newCtx);
 
         emit IndexUnsubscribed(token, publisher, indexId, subscriber, userData);
         emit SubscriptionRevoked(token, subscriber, publisher, indexId, userData);
@@ -882,7 +889,7 @@ contract InstantDistributionAgreementV1 is
 
         if (pendingDistribution > 0) {
             cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_UPDATED_NOOP;
-            vars.cbdata = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
+            (vars.cbdata, vars.callbackGasLimit) = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
             int256 signedPendingDistribution = pendingDistribution.toInt256();
 
             // adjust publisher's deposits
@@ -898,7 +905,8 @@ contract InstantDistributionAgreementV1 is
             emit SubscriptionDistributionClaimed(token, subscriber, publisher, indexId, pendingDistribution);
 
             cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_UPDATED_NOOP;
-            (, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+            (, newCtx) = AgreementLibrary.callAppAfterCallback(
+                cbStates, vars.cbdata, vars.callbackGasLimit, newCtx);
         } else {
             // nothing to be recorded in this case
             newCtx = ctx;